# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple inventory and transactions API",
        "title": "Inventory Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/analytics/inventory/low-stock": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of items that are below a specified stock threshold.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get Low Stock Items",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stock quantity threshold (default: system-defined)",
                        "name": "threshold",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of items to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Low stock items retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.GlobalResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.LowStockItemsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            }
        },
        "/analytics/inventory/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a summary of inventory statistics, including total items, total stock value, and average item value.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get Inventory Summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "As of date (YYYY-MM-DD)",
                        "name": "as_of",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory summary retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.GlobalResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.InventorySummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            }
        },
        "/analytics/revenue/growth": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve monthly revenue growth details including growth percentage, target achievement, burn rate, and top-selling items.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get Monthly Revenue Growth",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Month number (1-12)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Year (e.g., 2025)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Revenue growth target percentage",
                        "name": "target",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Monthly burn rate amount",
                        "name": "burn_rate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Revenue growth data retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.GlobalResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.MonthlyRevenueGrowthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            }
        },
        "/analytics/sales/trends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve sales trends for a given date range and interval (daily, weekly, or monthly).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get Sales Trends",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Time interval (daily|weekly|monthly)",
                        "name": "interval",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales trends retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.GlobalResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SalesTrendsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid parameters or date format",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "description": "Authenticate user credentials and return a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login an existing user",
                "parameters": [
                    {
                        "description": "SignIn Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Create a new user account with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Signup Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.GlobalResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.GlobalResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.InventorySummaryResponse": {
            "type": "object",
            "properties": {
                "as_of": {
                    "type": "string"
                },
                "average_item_value": {
                    "type": "number"
                },
                "total_available_qty": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_stock_value": {
                    "type": "number"
                }
            }
        },
...
        "dto.LowStockItemDetails": {
            "type": "object",
            "properties": {
                "available_qty": {
                    "type": "integer"
                },
                "item_name": {
                    "type": "string"
                }
            }
        },
        "dto.LowStockItemsResponse": {
            "type": "object",
            "properties": {
                "low_stock_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.LowStockItemDetails"
                    }
                },
                "threshold": {
                    "type": "integer"
                }
            }
        },
        "dto.MonthlyRevenueGrowthResponse": {
            "type": "object",
            "properties": {
                "burn_rate": {
                    "type": "number"
                },
                "growth_percent": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                },
                "previous_revenue": {
                    "type": "number"
                },
                "runway_months": {
                    "type": "number"
                },
                "target": {
                    "type": "number"
                },
                "target_achieved": {
                    "type": "boolean"
                },
                "top_selling_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TopSellingItemResponse"
                    }
                },
                "total_revenue": {
                    "type": "number"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "dto.SalesTrendRow": {
            "type": "object",
            "properties": {
                "period": {
                    "type": "string"
                },
                "revenue": {
                    "type": "number"
                },
                "units_sold": {
                    "type": "integer"
                }
            }
        },
        "dto.SalesTrendsResponse": {
            "type": "object",
            "properties": {
                "interval": {
                    "type": "string"
                },
                "trends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SalesTrendRow"
                    }
                }
            }
        },
        "dto.SignInRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                }
            }
        },
        "dto.SignupRequest": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 100
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                }
            }
        },
        "dto.TopSellingItemResponse": {
            "type": "object",
            "properties": {
                "item_name": {
                    "type": "string"
                },
                "revenue": {
                    "type": "number"
                },
                "total_sold": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}